MOVE] util 폴더링 해제
domain 제외 어떤 기능들이 어떻게 추가될지 모르므로 그냥 src 밑에 해당 기능별로 폴더링하는 것으로. prisma, swagger 같은 것들을 util이라고 보기엔 애매하니까
